refactor(global): restructure module
