+++ invenio_rdm_pure/converter/marc21_record/__init__.py
@@ -5,6 +5,13 @@
 # invenio-rdm-pure is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""Invenio module that adds pure."""
+"""Module for Marc21Record conversion."""
 
-__all__ = ("__version__", "InvenioRdmPure")
+from .marc21_record import ControlField, DataField, Marc21Record, SubField
+
+__all__ = (
+    "Marc21Record",
+    "ControlField",
+    "DataField",
+    "SubField",
+)
