--- invenio_rdm_pure/source/rdm/run/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright (C) 2021 Technische Universität Graz.
-#
-# invenio-rdm-pure is free software; you can redistribute it and/or modify it
-# under the terms of the MIT License; see LICENSE file for more details.
-
-"""Invenio module that adds pure."""
-
-__all__ = ("__version__", "InvenioRdmPure")
